Added remove flag and better profile check.

+++ install.sh
@@ -1,9 +1,19 @@
 echo "Installing BetterGit"
 
 PROFILE="$HOME/.bash_profile"
-WRITE_FLAG=false
-BETTER_CLONE_FUNC="function better_git_clone {sh C:\BetterGit\better_clone.sh \$args}"
+BETTER_CLONE_FUNC="{sh C:\BetterGit\better_clone.sh \$args}"
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+RM_FLAG=false
+while getopts "Rr" opt; do
+    case $opt in
+        r|R)
+            RM_FLAG=true
+            ;;
+        \?)
+            echo "Invalid option: -$OPTARG" >&2
+            ;;
+    esac
+done
 if [ -n "$1" ]; then
     PROFILE=$1
 fi
@@ -20,23 +30,22 @@ fi
 
 echo "Updating $INSTALL_DEST"
 cp "$SCRIPT_DIR/better_clone.sh" "$INSTALL_DEST\better_clone.sh"
-echo "$SCRIPT_DIR"
-rm -rf "$SCRIPT_DIR"
-echo "Cleaned up install files $(dirname "$0")"
+
+if [ "$RM_FLAG" = true ]; then
+    rm -rf "$SCRIPT_DIR"
+    echo "Cleaned up install files $(dirname "$0")"
+fi
+
 if [ -z "$PROFILE" ]; then
     echo "No profile found. Creating one at $PROFILE"
     touch $PROFILE
-    WRITE_FLAG=true
-else 
-    while IFS='' read -r line || [[ -n "$line" ]]; do
-        if [[ "$line" == *"better_git_clone"* ]]; then
-            echo "BetterGit already installed"
-            exit 0
-        fi
-    done < "$PROFILE"
-    WRITE_FLAG=true
 fi
-if [ "$WRITE_FLAG" = true ]; then
-    echo "Writing function to $PROFILE"
-    echo "$BETTER_CLONE_FUNC" >> $PROFILE
+if ! grep -q "function better_git_clone" "$PROFILE"; then
+    echo "function better_git_clone $BETTER_CLONE_FUNC" >> "$PROFILE"
+    echo "Added better_git_clone function to $PROFILE"
+fi
+
+if ! grep -q "function bgc" "$PROFILE"; then
+    echo "function bgc $BGC_FUNC" >> "$PROFILE"
+    echo "Added bgc function to $PROFILE"
 fi
